+++ client/web/src/insights/pages/dashboards/dashboard-page/components/dashboards-content/DashboardsContent.tsx
@@ -1,20 +1,16 @@
 import classnames from 'classnames'
 import MapSearchIcon from 'mdi-react/MapSearchIcon'
-import React, { useContext, useMemo, useState } from 'react'
+import React, { useRef, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 
-import { LoadingSpinner } from '@sourcegraph/react-loading-spinner'
 import { ExtensionsControllerProps } from '@sourcegraph/shared/src/extensions/controller'
 import { PlatformContextProps } from '@sourcegraph/shared/src/platform/context'
 import { SettingsCascadeProps } from '@sourcegraph/shared/src/settings/settings'
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
-import { useObservable } from '@sourcegraph/shared/src/util/useObservable'
 
 import { HeroPage } from '../../../../../../components/HeroPage'
 import { Settings } from '../../../../../../schema/settings.schema'
-import { InsightsViewGrid } from '../../../../../components'
-import { InsightsApiContext } from '../../../../../core/backend/api-provider'
-import { InsightDashboard, isVirtualDashboard } from '../../../../../core/types'
+import { isVirtualDashboard } from '../../../../../core/types'
 import { isSettingsBasedInsightsDashboard } from '../../../../../core/types/dashboard/real-dashboard'
 import { useDashboards } from '../../../../../hooks/use-dashboards/use-dashboards'
 import { AddInsightModal } from '../add-insight-modal/AddInsightModal'
@@ -22,7 +18,11 @@ import { DashboardMenu, DashboardMenuAction } from '../dashboard-menu/DashboardM
 import { DashboardSelect } from '../dashboard-select/DashboardSelect'
 import { DeleteDashboardModal } from '../delete-dashboard-modal/DeleteDashboardModal'
 
+import { DashboardInsights } from './components/dashboard-inisghts/DashboardInsights'
 import styles from './DashboardsContent.module.scss'
+import { useCopyURLHandler } from './hooks/use-copy-url-handler'
+import { useDashboardSelectHandler } from './hooks/use-dashboard-select-handler'
+import { findDashboardByUrlId } from './utils/find-dashboard-by-url-id'
 import { isDashboardConfigurable } from './utils/is-dashboard-configurable'
 
 export interface DashboardsContentProps
@@ -49,36 +49,10 @@ export const DashboardsContent: React.FunctionComponent<DashboardsContentProps> 
     const [isAddInsightOpen, setAddInsightsState] = useState<boolean>(false)
     const [isDeleteDashboardActive, setDeleteDashboardActive] = useState<boolean>(false)
 
-    const currentDashboard = dashboards.find(dashboard => {
-        if (isVirtualDashboard(dashboard)) {
-            return (
-                dashboard.id === dashboardID.toLowerCase() || dashboard.type.toLowerCase() === dashboardID.toLowerCase()
-            )
-        }
-
-        return (
-            dashboard.id === dashboardID ||
-            dashboard.title.toLowerCase() === dashboardID?.toLowerCase() ||
-            (isSettingsBasedInsightsDashboard(dashboard) &&
-                dashboard.settingsKey.toLowerCase() === dashboardID?.toLowerCase())
-        )
-    })
-
-    const handleDashboardSelect = (dashboard: InsightDashboard): void => {
-        if (isVirtualDashboard(dashboard)) {
-            history.push(`/insights/dashboards/${dashboard.type}`)
-
-            return
-        }
-
-        if (isSettingsBasedInsightsDashboard(dashboard)) {
-            history.push(`/insights/dashboards/${dashboard.settingsKey}`)
-
-            return
-        }
-
-        history.push(`/insights/dashboards/${dashboard.id}`)
-    }
+    const currentDashboard = findDashboardByUrlId(dashboards, dashboardID)
+    const handleDashboardSelect = useDashboardSelectHandler()
+    const [copyURL, isCopied] = useCopyURLHandler()
+    const menuReference = useRef<HTMLButtonElement | null>(null)
 
     const handleSelect = (action: DashboardMenuAction): void => {
         switch (action) {
@@ -86,23 +60,28 @@ export const DashboardsContent: React.FunctionComponent<DashboardsContentProps> 
                 if (!isVirtualDashboard(currentDashboard) && isSettingsBasedInsightsDashboard(currentDashboard)) {
                     history.push(`/insights/dashboards/${currentDashboard.settingsKey}/edit`)
                 }
-
                 return
             }
-
             case DashboardMenuAction.AddRemoveInsights: {
                 setAddInsightsState(true)
-
                 return
             }
-
             case DashboardMenuAction.Delete: {
                 setDeleteDashboardActive(true)
+                return
+            }
+            case DashboardMenuAction.CopyLink: {
+                copyURL()
+
+                // Re-trigger trigger tooltip event catching logic to activate
+                // copied tooltip appearance
+                requestAnimationFrame(() => {
+                    menuReference.current?.blur()
+                    menuReference.current?.focus()
+                })
 
                 return
             }
-
-            // Implement other actions
         }
     }
 
@@ -118,7 +97,13 @@ export const DashboardsContent: React.FunctionComponent<DashboardsContentProps> 
                     className={classnames(styles.dashboardSelect, 'mr-2')}
                 />
 
-                <DashboardMenu dashboard={currentDashboard} settingsCascade={settingsCascade} onSelect={handleSelect} />
+                <DashboardMenu
+                    innerRef={menuReference}
+                    tooltipText={isCopied ? 'Copied!' : undefined}
+                    dashboard={currentDashboard}
+                    settingsCascade={settingsCascade}
+                    onSelect={handleSelect}
+                />
             </section>
 
             <hr className="mt-2 mb-3" />
@@ -148,41 +133,6 @@ export const DashboardsContent: React.FunctionComponent<DashboardsContentProps> 
                     platformContext={platformContext}
                     onClose={() => setDeleteDashboardActive(false)}
                 />
-            )}
-        </div>
-    )
-}
-
-interface DashboardInsightsProps extends ExtensionsControllerProps, TelemetryProps {
-    /**
-     * Dashboard specific insight ids.
-     */
-    insightIds?: string[]
-}
-
-/**
- * Renders code insight view grid.
- */
-const DashboardInsights: React.FunctionComponent<DashboardInsightsProps> = props => {
-    const { telemetryService, extensionsController, insightIds } = props
-    const { getInsightCombinedViews } = useContext(InsightsApiContext)
-
-    const views = useObservable(
-        useMemo(() => getInsightCombinedViews(extensionsController?.extHostAPI, insightIds), [
-            insightIds,
-            extensionsController,
-            getInsightCombinedViews,
-        ])
-    )
-
-    return (
-        <div>
-            {views === undefined ? (
-                <div className="d-flex w-100">
-                    <LoadingSpinner className="my-4" />
-                </div>
-            ) : (
-                <InsightsViewGrid views={views} hasContextMenu={true} telemetryService={telemetryService} />
             )}
         </div>
     )